🎨 change popular components to main page
